requirements.txt
python-dotenv == 1.0.0
psycopg2-binary == 2.9.6
sqlalchemy == 2.0.19

